PM_Monthly_Activity {rsellTables}	R Documentation

Creates a Poshmark Monthly Activity Table

Description

This function creates a Poshmark Monthly Activity Table from the Inventory Table.

Usage

PM_Monthly_Activity(pm_sales, flt_year, flt_month)

Arguments

pm_sales
The Poshmark Sales Table for the user.

Details

Requires: dplyr, lubridate

Value

A Poshmark Monthly Activity Table as a data frame.

[Package rsellTables version 1.0.0 Index]