Source_Inventory_Summary {rsellTables}	R Documentation

Creates a Source Inventory Summary Table

Description

This function creates a Source Inventory Summary Table from the Inventory Table.

Usage

Source_Inventory_Summary(inv_table)

Arguments

inv_table
The Inventory Table for the user.

Details

Requires: dplyr

Value

A Source Inventory Summary Table as a data frame.

[Package rsellTables version 1.0.0 Index]